+++ src/instagram.nim
@@ -4,6 +4,8 @@ export core
 # API
 import instagram/api/get
 export get
+import instagram/api/post
+export post
 
 proc hasNextPage*(obj: IgFollowersAndFollowing or IgFeed): bool =
   ## Check if exists next page