+++ src/instagram/api/get.nim
@@ -1,3 +1,5 @@
+## POST actions in Instagram internal API
+
 import std/asyncdispatch
 from std/httpcore import HttpGet, HttpPost